Add health notices feature with real-time WebSocket integration
- Add health notices screen with filtering by severity (fatal/critical/warning/notice)
  and device type (AP/Switch/ONT)
- Add health gauge widget to dashboard showing critical issue counts
- Integrate health notices from device data via WebSocket cache
- Add severity counting extension to DRY up duplicate counting logic
- Wrap debug prints in kDebugMode for production builds
- Fix unused_result warnings by using ref.invalidate() instead of ref.refresh()
- Apply 43 lint fixes across 27 files via dart fix

+++ lib/features/devices/presentation/widgets/image_viewer_dialog.dart
@@ -69,9 +69,9 @@ class _ImageViewerDialogState extends State<ImageViewerDialog> {
   Widget build(BuildContext context) {
     final media = MediaQuery.of(context);
     final cacheWidth =
-        (media.size.width * media.devicePixelRatio).round().clamp(1, 4096).toInt();
+        (media.size.width * media.devicePixelRatio).round().clamp(1, 4096);
     final cacheHeight =
-        (media.size.height * media.devicePixelRatio).round().clamp(1, 4096).toInt();
+        (media.size.height * media.devicePixelRatio).round().clamp(1, 4096);
 
     return Scaffold(
       backgroundColor: Colors.transparent,
